move MealHeader to component package
